replace js implementation completely
