+++ src/core/save/G3SAV.ts
@@ -179,9 +179,9 @@ export class G3SaveBackup {
         const buffer = this.pcDataContiguous.slice(4 + i * 80, 4 + (i + 1) * 80).buffer
         box.boxSlots[slot] = PK3.fromSlotBytes(buffer)
       } catch (e) {
-        // console.error(
-        //   `File has invalid Pokémon data at box ${Math.floor(i / 30)}/slot ${slot}: ${e}`
-        // )
+        console.error(
+          `File has invalid Pokémon data at box ${Math.floor(i / 30)}/slot ${slot}: ${e}`
+        )
       }
     }
